Initial controller implementation complete

+++ src/main/java/com/revature/controllers/RoomController.java
@@ -22,8 +22,11 @@ public class RoomController {
     RoomService roomService;
 
     @PostMapping("/locations/{locationId}/buildings/{buildingId}/rooms")
-    public ResponseEntity<RoomDto> createRoom(@RequestBody RoomDto roomDto, @RequestHeader(name = "Authorization") String auth){
-        if(auth.equals("Authorized")) {
+    public ResponseEntity<RoomDto> createRoom(@RequestBody RoomDto roomDto, @RequestHeader(name = "Authorization", required = false) String auth){
+
+        if(auth == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        } else if (auth.equals("Authorized")) {
             Room room = new Room(roomDto);
             room = this.roomService.createRoom(room);
             return ResponseEntity.status(HttpStatus.CREATED).body(new RoomDto(room));
@@ -33,40 +36,62 @@ public class RoomController {
     }
 
     @GetMapping("/locations/{locationId}/buildings/{buildingId}/rooms/{roomId}")
-    public ResponseEntity<Object> getRoomById(@PathVariable int locationId, @PathVariable int buildingId, @PathVariable int roomId){
+    public ResponseEntity<RoomDto> getRoomById(@PathVariable int locationId, @PathVariable int buildingId, @PathVariable int roomId, @RequestHeader(name = "Authorization", required = false) String auth){
+
         try{
-            Room room = this.roomService.getRoomById(roomId);
-            return ResponseEntity.status(HttpStatus.OK).body(new RoomDto(room));
+            if (auth.equals("Authorized")) {
+                Room room = this.roomService.getRoomById(roomId);
+                return ResponseEntity.status(HttpStatus.OK).body(new RoomDto(room));
+            } else {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
         }catch(RoomNotFoundException e){
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Error message: Room not found");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
     }
 
     @GetMapping("/locations/{locationId}/buildings/{buildingId}/rooms")
-    public ResponseEntity<List<RoomDto>> getAllRooms(@PathVariable int locationId, @PathVariable int buildingId){
-        List<Room> rooms = this.roomService.getAllRooms();
-        List<RoomDto> roomDtos = new ArrayList<>();
-        for(Room r : rooms){
-            roomDtos.add(new RoomDto(r));
+    public ResponseEntity<List<RoomDto>> getAllRooms(@PathVariable int locationId, @PathVariable int buildingId,@RequestHeader(name = "Authorization", required = false) String auth){
+        if (auth.equals("Authorized")) {
+            List<Room> rooms = this.roomService.getAllRooms();
+            List<RoomDto> roomDtos = new ArrayList<>();
+            for (Room r : rooms) {
+                roomDtos.add(new RoomDto(r));
+            }
+            return ResponseEntity.status(HttpStatus.OK).body(roomDtos);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
-        return ResponseEntity.status(HttpStatus.OK).body(roomDtos);
     }
 
     @PutMapping("/locations/{locationId}/buildings/{buildingId}/rooms/{roomId}")
-    public ResponseEntity<RoomDto> updateRoom(@PathVariable int locationId, @PathVariable int buildingId,@PathVariable int roomId, @RequestBody RoomDto roomDto){
-        try{
-            Room room = new Room(roomDto);
-            room = this.roomService.updateRoom(room);
-            return ResponseEntity.status(HttpStatus.OK).body(new RoomDto(room));
+    public ResponseEntity<RoomDto> updateRoom(@PathVariable int locationId, @PathVariable int buildingId,@PathVariable int roomId, @RequestBody RoomDto roomDto, @RequestHeader(name = "Authorization", required = false) String auth){
+        try {
+            if (auth == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+            }else if ( auth.equals("Authorized")) {
+                Room room = new Room(roomDto);
+                room = this.roomService.updateRoom(room);
+                return ResponseEntity.status(HttpStatus.OK).body(new RoomDto(room));
+            } else {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
         }catch(RoomNotFoundException e){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
     }
 
     @DeleteMapping("/locations/{locationId}/buildings/{buildingId}/rooms/{roomId}")
-    public ResponseEntity<Boolean> deleteRoom(@PathVariable int locationId, @PathVariable int buildingId, @PathVariable int roomId){
-        boolean result = this.roomService.deleteRoom(roomId);
-        return ResponseEntity.status(HttpStatus.OK).body(result);
+    public ResponseEntity<Boolean> deleteRoom(@PathVariable int locationId, @PathVariable int buildingId, @PathVariable int roomId, @RequestHeader(name = "Authorization", required = false) String auth){
+        if(auth == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        } else if (auth.equals("Authorized")) {
+            boolean result = this.roomService.deleteRoom(roomId);
+            return ResponseEntity.status(HttpStatus.OK).body(result);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
     }
 
 }
